fix some code smells
