+++ src/consts/const-addimages.js
@@ -1,0 +1,17 @@
+export const PRODUCT_PHOTO_TEXT = {
+  PRIMARY: 'Головне фото',
+  ADDITIONAL: 'Додаткові фото'
+};
+
+export const IMAGES_INDEXES = {
+  FIRST_ADDITIONAL_IMAGE: 0,
+  SECOND_ADDITIONAL_IMAGE: 1,
+  THIRD_ADDITIONAL_IMAGE: 2
+};
+
+export const imageUploadInputsId = {
+  mainImageInput: 'mainImageInput',
+  imageInput1: 'ImgInput1',
+  imageInput2: 'ImgInput2',
+  imageInput3: 'ImgInput3'
+};
